error tratament

+++ mercado-livro/src/main/kotlin/com/mercadolivro/controller/response/CreateBookResponse.kt
@@ -1,13 +1,12 @@
 package com.mercadolivro.controller.response
 
 import com.mercadolivro.enums.BookStatus
-import com.mercadolivro.model.Book
-import java.util.UUID
+import com.mercadolivro.entitys.Book
 
 class CreateBookResponse(
     val name: String,
     val status: BookStatus,
-    val price: Int,
+    val price: Double,
     val image: String,
     val customerId: Int
 )
